New file for Nav tests & declared leftClick once
